@@ -1,5 +1,5 @@
 /*
- * Copyright 2009-2011 The VOTCA Development Team (http://www.votca.org)
+ * Copyright 2009-2019 The VOTCA Development Team (http://www.votca.org)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,13 +16,13 @@
  */
 
 #ifndef _TRAJ_FORCE_H
-#define	_TRAJ_FORCE_H
+#define _TRAJ_FORCE_H
 
-#include <votca/tools/property.h>
 #include <boost/numeric/ublas/vector.hpp>
 #include <votca/csg/csgapplication.h>
 #include <votca/csg/trajectoryreader.h>
 #include <votca/csg/trajectorywriter.h>
+#include <votca/tools/property.h>
 
 using namespace votca::csg;
 using namespace std;
@@ -31,27 +31,25 @@ using namespace std;
    \brief Adds/subtracts forces from given atomistic trajectories
 **/
 
-class TrajForce
-: public CsgApplication
-{
+class TrajForce : public CsgApplication {
  public:
-  string ProgramName() { return "traj_force"; }
-  void HelpText(ostream &out) {
+  string ProgramName() override { return "traj_force"; }
+  void HelpText(ostream &out) override {
     out << "Adds/subtracts forces from given atomistic trajectories";
   }
 
-  bool DoTrajectory() {return true;}
-  bool DoMapping() {return false;}
+  bool DoTrajectory() override { return true; }
+  bool DoMapping() override { return false; }
 
-  void Initialize(void);
-  bool EvaluateOptions();
+  void Initialize(void) override;
+  bool EvaluateOptions() override;
 
   /// \brief called before the first frame
-  void BeginEvaluate(Topology *top, Topology *top_atom);
+  void BeginEvaluate(Topology *top, Topology *top_atom) override;
   /// \brief called after the last frame
-  void EndEvaluate();
+  void EndEvaluate() override;
   /// \brief called for each frame which is mapped
-  void EvalConfiguration(Topology *conf, Topology *conf_atom);
+  void EvalConfiguration(Topology *conf, Topology *conf_atom) override;
 
  protected:
   /// \brief Scaling of forces, +1 for addition and -1 for subtraction
@@ -65,5 +63,4 @@ class TrajForce
   TrajectoryWriter *_trjwriter;
 };
 
-#endif	/* _TRAJ_FORCE_H */
-
+#endif /* _TRAJ_FORCE_H */
